Moved SignedString.cs into ServerComms
