+++ RightToAskClient/RightToAskClient/RightToAskClient/CryptoUtils/SignatureVerificationService.cs
@@ -5,6 +5,7 @@ using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Crypto.Signers;
 using RightToAskClient.HttpClients;
 using RightToAskClient.Models;
+using RightToAskClient.Models.ServerCommsData;
 
 namespace RightToAskClient.CryptoUtils
 {
